# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive development toolkit for Odoo environments with profiling, monitoring, and optimization features.",
        "title": "OdooDevTools API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://example.com/support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/agent/errors": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Agent endpoint: receive a batch of Odoo error/exception events",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent"
                ],
                "summary": "Ingest error events",
                "parameters": [
                    {
                        "description": "Error batch",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.IngestErrorsRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/agent/schema": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Agent endpoint: persist a collected Odoo schema snapshot",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent"
                ],
                "summary": "Store schema snapshot",
                "parameters": [
                    {
                        "description": "Schema payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.StoreSchemaRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.SchemaSnapshotResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
...
        "/environments/{env_id}/schema": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return schema snapshot summaries for the given environment, newest first",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schema"
                ],
                "summary": "List schema snapshots",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment ID (UUID)",
                        "name": "env_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SchemaSnapshotListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/environments/{env_id}/schema/latest": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return the most recently captured schema snapshot for the given environment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schema"
                ],
                "summary": "Get latest schema snapshot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment ID (UUID)",
                        "name": "env_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SchemaSnapshotResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/environments/{env_id}/schema/models": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Filter models in the latest snapshot by technical name or display label",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schema"
                ],
                "summary": "Search schema models",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment ID (UUID)",
                        "name": "env_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search term (substring, case-insensitive)",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SearchModelsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/api.ErrorCode"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ErrorDetail"
                    }
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "api.ErrorCode": {
            "type": "string",
            "enum": [
                "VALIDATION_ERROR",
                "UNAUTHORIZED",
                "FORBIDDEN",
                "NOT_FOUND",
                "CONFLICT",
                "UNPROCESSABLE_ENTITY",
                "RATE_LIMITED",
                "BAD_REQUEST",
                "PAYLOAD_TOO_LARGE",
                "INTERNAL_ERROR",
                "SERVICE_UNAVAILABLE",
                "TIMEOUT",
                "DATABASE_ERROR",
                "TENANT_NOT_FOUND",
                "ENVIRONMENT_NOT_FOUND",
                "AGENT_NOT_CONNECTED",
                "AGENT_NOT_FOUND",
                "QUOTA_EXCEEDED",
                "PLAN_LIMIT_REACHED",
                "INVALID_API_KEY",
                "EXPIRED_API_KEY",
                "INVALID_TOKEN",
                "EXPIRED_TOKEN",
                "INVALID_CREDENTIALS",
                "EMAIL_ALREADY_EXISTS",
                "SLUG_ALREADY_EXISTS",
                "INVALID_SCOPE",
                "SCHEMA_SNAPSHOT_NOT_FOUND",
                "MIGRATION_SCAN_NOT_FOUND",
                "ERROR_GROUP_NOT_FOUND",
                "ALERT_RULE_NOT_FOUND",
                "ALERT_CHANNEL_NOT_FOUND",
                "PROFILER_RECORDING_NOT_FOUND",
                "ANON_PROFILE_NOT_FOUND",
                "ANON_JOB_NOT_FOUND",
                "INVALID_ODOO_VERSION",
                "INVALID_DOMAIN"
            ],
            "x-enum-comments": {
                "ErrCodeExpiredAPIKey": "nolint:gosec // not a credential",
                "ErrCodeInvalidAPIKey": "nolint:gosec // not a credential",
                "ErrCodeInvalidCredentials": "nolint:gosec // not a credential"
            },
            "x-enum-descriptions": [
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "nolint:gosec // not a credential",
                "nolint:gosec // not a credential",
                "",
                "",
                "nolint:gosec // not a credential",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                ""
            ],
            "x-enum-varnames": [
                "ErrCodeValidation",
                "ErrCodeUnauthorized",
                "ErrCodeForbidden",
                "ErrCodeNotFound",
                "ErrCodeConflict",
                "ErrCodeUnprocessable",
                "ErrCodeRateLimited",
                "ErrCodeBadRequest",
                "ErrCodePayloadTooLarge",
                "ErrCodeInternal",
                "ErrCodeUnavailable",
                "ErrCodeTimeout",
                "ErrCodeDatabase",
                "ErrCodeTenantNotFound",
                "ErrCodeEnvironmentNotFound",
                "ErrCodeAgentNotConnected",
                "ErrCodeAgentNotFound",
                "ErrCodeQuotaExceeded",
                "ErrCodePlanLimitReached",
                "ErrCodeInvalidAPIKey",
                "ErrCodeExpiredAPIKey",
                "ErrCodeInvalidToken",
                "ErrCodeExpiredToken",
                "ErrCodeInvalidCredentials",
                "ErrCodeEmailAlreadyExists",
                "ErrCodeSlugAlreadyExists",
                "ErrCodeInvalidScope",
                "ErrCodeSchemaSnapshotNotFound",
                "ErrCodeMigrationScanNotFound",
                "ErrCodeErrorGroupNotFound",
                "ErrCodeAlertRuleNotFound",
                "ErrCodeAlertChannelNotFound",
                "ErrCodeProfilerRecordingNotFound",
                "ErrCodeAnonProfileNotFound",
                "ErrCodeAnonJobNotFound",
                "ErrCodeInvalidOdooVersion",
                "ErrCodeInvalidDomain"
            ]
        },
        "api.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "value": {}
            }
        },
...
        "dto.ErrorEventPayload": {
            "type": "object",
            "required": [
                "captured_at",
                "error_type",
                "message",
                "signature"
            ],
            "properties": {
                "affected_uid": {
                    "type": "integer"
                },
                "captured_at": {
                    "type": "string"
                },
                "error_type": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "module": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "traceback": {
                    "type": "string"
                }
            }
        },
...
        "dto.IngestErrorsRequest": {
            "type": "object",
            "required": [
                "env_id",
                "events"
            ],
            "properties": {
                "env_id": {
                    "type": "string"
                },
                "events": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/dto.ErrorEventPayload"
                    }
                },
                "spike_threshold": {
                    "type": "integer"
                }
            }
        },
...
        "dto.SchemaSnapshotListItem": {
            "type": "object",
            "properties": {
                "captured_at": {
                    "type": "string"
                },
                "diff_ref": {
                    "type": "string"
                },
                "env_id": {
                    "type": "string"
                },
                "field_count": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "model_count": {
                    "type": "integer"
                }
            }
        },
        "dto.SchemaSnapshotListResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "snapshots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SchemaSnapshotListItem"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.SchemaSnapshotResponse": {
            "type": "object",
            "properties": {
                "acl_rules": {
                    "type": "object"
                },
                "captured_at": {
                    "type": "string"
                },
                "diff_ref": {
                    "type": "string"
                },
                "env_id": {
                    "type": "string"
                },
                "field_count": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "model_count": {
                    "type": "integer"
                },
                "models": {
                    "type": "object"
                },
                "record_rules": {
                    "type": "object"
                }
            }
        },
        "dto.SearchModelsResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "models": {
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.StoreSchemaRequest": {
            "type": "object",
            "required": [
                "acl_rules",
                "env_id",
                "models",
                "record_rules"
            ],
            "properties": {
                "acl_rules": {
                    "type": "object"
                },
                "env_id": {
                    "type": "string"
                },
                "field_count": {
                    "type": "integer"
                },
                "model_count": {
                    "type": "integer"
                },
                "models": {
                    "type": "object"
                },
                "record_rules": {
                    "type": "object"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"ApiKey\" followed by a space and your API key (e.g. \"ApiKey odt_xxx...\").",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}